Refactor hermit-abi for in kernel use
This separates the linked operating system functions from other structs
or constants and hides them when the `in-kernel` feature is active.
This allows hermit-abi to be included in the `Cargo.toml` file of
`libhermit-rs` without linking to the missing functions that would be
exposed by `hermit-sys`.
